+++ syndicate/core/__init__.py
@@ -187,11 +187,14 @@ def initialize_project_state():
     from syndicate.core.project_state.sync_processor import sync_project_state
     global PROJECT_STATE
     if not ProjectState.check_if_project_state_exists(CONFIG.project_path):
-        USER_LOG.warn("Config is set and generated but project "
-                      "state does not exist, seems that you've come from the "
-                      "previous version.")
-        USER_LOG.warn("Generating project state file "
-                      "(.syndicate) from the existing structure...")
+        USER_LOG.warn(
+            "Config is set and generated, but project state file does not "
+            "exist."
+        )
+        USER_LOG.warn(
+            "Generating project state file '.syndicate' from the existing "
+            "structure..."
+        )
         PROJECT_STATE = ProjectState.build_from_structure(CONFIG)
     else:
         PROJECT_STATE = ProjectState(project_path=CONFIG.project_path)